Refactoring des doubles et des fonction utilisant des doubles pour utiliser des Floats

+++ EnVoiture/CarWidget.cs
@@ -18,7 +18,7 @@ namespace EnVoiture
         /// <param name="rectangle">Rectangle sur lequel baser la géométrie de la voiture</param>
         public CarWidget(Rectangle rectangle)
         {
-            this.Car = new Car(rectangle,0.0);
+            this.Car = new Car(rectangle,0.0F);
         }
 
         /// <summary>
@@ -28,7 +28,7 @@ namespace EnVoiture
         /// <param name="y">Position y du haut</param>
         /// <param name="width">Largeur</param>
         /// <param name="height">Hauteur</param>
-        public CarWidget(int x, int y, int width, int height, double vMax)
+        public CarWidget(int x, int y, int width, int height, float vMax)
         {
             this.Car = new Car(x, y, width, height,vMax);
         }
